translator: DWARF: Correct handling of passed structures
Passed structures are handled differently for return values and
arguments: for return values, an address to a stack location is placed
in v0, for arguments, the first four words are passed in a0...a3,
while the rest is passed on the stack.

Return -1 for aggregate types for functions (better safe than sorry).