libs: Workaround the need for linking with --whole-archive
Some of the AR archives are actually built as ELF object files, to
allow global constructors to run even if nothing in the file is used.
For files which do contain exported symbols, these are declared WEAK
to allow double linking (which e.g., cmake likes to do).

In libraries which contain more than one .c-file, the files are simply
included into one of them.